fix: Unapprove surplus token in swap
The current swap logic sets exact approval amounts for the Sushi router.
For consistence, if not all token was consumed by the swap, the approval
should then be reset to 0.

An alternative would be to just set a max. approval for the sushi router,
which would probably be the better solution here. But that's open for
discussion :)