feat: Add support for fixed input token offsetting to OffsetHelper
Added two new overloads to the public functions `autoOffsetUsingToken`
and `autoOffsetUsingETH` that deal with a fixed input amount.
The resulting redeemed TCO2 is thus variable and dependent on the
Pool token exchange rate at the time of the call.

The corresponding overloads were implemented on the `swap` function.

The estimated amount of offset TCO0 can be queried with the new function
`calculateExpectedPoolTokenAmount`, which has two overloads for sending
token or MATIC.