feat: Simplify swap functions
3 out of 4 `swap` function overloads could be simplified by just passing
all ETH or token to the sushi router for swapping, instead of
calculating the needed input or output amount first. Since all logic
is happening within the same transaction, calculating the amounts
first or just executing the swap directly leads to the same behavior,
but has simplified logic.

The `swap` function that swaps variable erc20 token for a fixed offset
couldn't be simplified in this way because the amount of input token
needs to be known to transfer it to the OffsetHelper first, before
executing the swap. For this function to also get simplified, the amount
would need to be passed as a function parameter, so I just kept the
current implementation as to not change the public API.