feat: Rename public overloaded functions to unique names
To disambiguate overloaded public functions, the `autoOffset...` functions
got renamed to
- `autoOffsetExactOutToken`
- `autoOffsetExactInToken`
- `autoOffsetExactOutETH`
- `autoOffsetExactInETH`
and the four `swap` overloads got renamed to
- `swapExactOutToken`
- `swapExactInToken`
- `swapExactOutETH`
- `swapExactInETH`

For consistency, function `autoOffsetUsingPoolToken` got renamed to
`autoOffsetPoolToken`, removing the `Using` like for the other
`autoOffset...` functions.

And finally, the overloads for function `calculateExpectedPoolTokenAmount`
were resolved by renaming them to `calculateExpectedPoolTokenForETH` and
`calculateExpectedPoolTokenForToken`.

The word `Exact` was used in favor if `Fixed` to be in line with the
Uniswap convention of naming their functions. The internal functions
`calculateFixed{In,Out}Swap` got renamed too to follow this convention.

The unused `hardhat/console.sol` imports were removed as well.