refactor(tests): use hardhat fixtures
Without hardhat fixtures, the code in `beforeEach` hook will
re-run before each test. This significantly slows down
the execution of tests.

With hardhat fixtures, the fixture is called once.
After, instead of being executed again, the state of the
network is reset to the point where it was right after
the fixture was first executed. This significantly increases
test execution speed (2-3x in this case).